# Some small performance optimizations